+++ CHANGELOG.md
@@ -5,7 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Video download button
+- Subtitle download button
 ### Changed
+- `README.md` updated
 - `service-worker.js` modified
 ## [0.8] - 2026-03-26
 ### Added